Added swiftui models, google books api, and some ui updates

+++ virtual-library/Code/Features/Discover/Views/RatingView.swift
@@ -9,13 +9,13 @@ import Foundation
 import SwiftUI
 
 struct RatingView: View {
-    var rating: Int
+    var rating: Double
     var body: some View {
         HStack(spacing: 4) {
             ForEach(1...5, id: \.self) { index in
                 Image(systemName: "star.fill")
                     .font(.caption2)
-                    .foregroundColor(index <= rating ? .yellow : .gray.opacity(0.5))
+                    .foregroundColor(Double(index) <= rating ? .yellow : .gray.opacity(0.5))
             }
             
             Text("(\(rating))")
